seminar_10_auth_in_ms/message-service/requirements.txt
fastapi
uvicorn[standard]
PyJWT
pydantic-settings